debug test

+++ src/main/java/sg/nus/iss/java/team7/repos/AdministratorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import sg.nus.iss.java.team7.models.Administrator;
 
-public interface AdministratorRepository extends JpaRepository<Administrator,Integer>{
+public interface AdministratorRepository extends JpaRepository<Administrator,Long>{
     
 }
